refactor(content): move notes into content/notes.ts
Notes were self-contained (NoteSchema, flat-directory walk, alias
support). The _allNotes prod-only singleton becomes a createProdMemo —
same dev-recompute semantics for HMR freshness, expressed through the
shared cache helper instead of a hand-rolled nullable. 5 consumers
update; the static-params unit test stubs the new module.
